Fix linux build

+++ fizzbuzz_server/timing.cc
@@ -8,7 +8,7 @@ TimeLog::TimeLog()
 	mEntries.reserve(MAX_TIME_LOG_SIZE);
 }
 
-long long ToEpoch(Timestamp ts)
+unsigned long long ToEpoch(Timestamp ts)
 {
 	return std::chrono::duration_cast<std::chrono::nanoseconds>(ts.time_since_epoch()).count();
 }
@@ -20,8 +20,8 @@ void TimeLog::DumpLog()
 	fprintf(f, "Header: TimeLogEntry: request=uint64,startTS=timestamp,finishParsingTS=timestamp,finishProcessingTS=timestamp,finishSendTS=timestamp\n");
 	for (auto& entry : mEntries)
 	{
-		fprintf(f, "Data: TimeLogEntry: request=%lld,startTS=%lld,finishParsingTS=%lld,finishProcessingTS=%lld,finishSendTS=%lld\n",
-			entry.request, ToEpoch(entry.startTS), ToEpoch(entry.finishParsingTS),
+		fprintf(f, "Data: TimeLogEntry: request=%llu,startTS=%llu,finishParsingTS=%llu,finishProcessingTS=%llu,finishSendTS=%llu\n",
+			(unsigned long long)entry.request, ToEpoch(entry.startTS), ToEpoch(entry.finishParsingTS),
 			ToEpoch(entry.finishProcessingTS), ToEpoch(entry.finishSendTS));
 	}
 	fclose(f);
